+++ src/Dibi/Drivers/Engines/OracleEngine.php
@@ -7,15 +7,17 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\Engines;
 
 use Dibi;
+use Dibi\Drivers\Connection;
+use Dibi\Drivers\Engine;
 
 
 /**
- * The reflector for ODBC connections.
+ * The reflector for Oracle database.
  */
-class OdbcReflector implements Engine
+class OracleEngine implements Engine
 {
 	public function __construct(
 		private readonly Connection $driver,
@@ -28,18 +30,17 @@ class OdbcReflector implements Engine
 	 */
 	public function getTables(): array
 	{
-		$res = odbc_tables($this->driver->getResource());
+		$res = $this->driver->query('SELECT * FROM cat');
 		$tables = [];
-		while ($row = odbc_fetch_array($res)) {
-			if ($row['TABLE_TYPE'] === 'TABLE' || $row['TABLE_TYPE'] === 'VIEW') {
+		while ($row = $res->fetch(false)) {
+			if ($row[1] === 'TABLE' || $row[1] === 'VIEW') {
 				$tables[] = [
-					'name' => $row['TABLE_NAME'],
-					'view' => $row['TABLE_TYPE'] === 'VIEW',
+					'name' => $row[0],
+					'view' => $row[1] === 'VIEW',
 				];
 			}
 		}
 
-		odbc_free_result($res);
 		return $tables;
 	}
 
@@ -49,22 +50,20 @@ class OdbcReflector implements Engine
 	 */
 	public function getColumns(string $table): array
 	{
-		$res = odbc_columns($this->driver->getResource());
+		$res = $this->driver->query('SELECT * FROM "ALL_TAB_COLUMNS" WHERE "TABLE_NAME" = ' . $this->driver->escapeText($table));
 		$columns = [];
-		while ($row = odbc_fetch_array($res)) {
-			if ($row['TABLE_NAME'] === $table) {
-				$columns[] = [
-					'name' => $row['COLUMN_NAME'],
-					'table' => $table,
-					'nativetype' => $row['TYPE_NAME'],
-					'size' => $row['COLUMN_SIZE'],
-					'nullable' => (bool) $row['NULLABLE'],
-					'default' => $row['COLUMN_DEF'],
-				];
-			}
+		while ($row = $res->fetch(true)) {
+			$columns[] = [
+				'table' => $row['TABLE_NAME'],
+				'name' => $row['COLUMN_NAME'],
+				'nativetype' => $row['DATA_TYPE'],
+				'size' => $row['DATA_LENGTH'] ?? null,
+				'nullable' => $row['NULLABLE'] === 'Y',
+				'default' => $row['DATA_DEFAULT'],
+				'vendor' => $row,
+			];
 		}
 
-		odbc_free_result($res);
 		return $columns;
 	}
