--- src/Dibi/Drivers/Sqlite3Driver.php
@@ -1,18 +1,0 @@
-<?php
-
-/**
- * This file is part of the Dibi, smart database abstraction layer (https://dibiphp.com)
- * Copyright (c) 2005 David Grudl (https://davidgrudl.com)
- */
-
-declare(strict_types=1);
-
-namespace Dibi\Drivers;
-
-
-/**
- * Alias for SqliteDriver driver.
- */
-class Sqlite3Driver extends SqliteDriver
-{
-}
